Fix weird permissions issue on .vim dir.
The vim-plug install script seems to need to be run as my user even
though I chown the entire $HOME.  If not things inside ~/.vim had a
unknown user.  Weird thing is that this only caused an issue when the
container was started on a Mac in boot2docker, Linux hosts seemed fine.
This might be a boot2docker oddity or due to the fact that the Linux
hosts were all on an older version of Docker whereas my Mac was on the
latest (1.6).  Anywhoo, this fixes things for both.